backend/requirements.txt
Flask==2.3.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
qrcode==7.4.2
pillow==10.0.0
twilio==8.10.0
openai==2.29.0
pytesseract==0.3.10
opencv-python-headless==4.8.0.74